requirements.txt
requests>=2.28.0
google-cloud-bigquery>=3.10.0
pandas-gbq>=0.19.0
python-dotenv>=1.0.0
pandas>=2.0.0
python-slugify>=8.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0